requirements.txt
asgiref==3.4.1
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.6
cryptography==3.4.8
defusedxml==0.7.1
diff-match-patch==20200713
Django==3.2.7
django-allauth==0.45.0
django-import-export==2.6.0
django-rest-auth==0.9.5
djangorestframework==3.12.4
et-xmlfile==1.1.0
greenlet==1.1.1
idna==3.2
jdcal==1.4.1
MarkupPy==1.14
numpy==1.21.2
oauthlib==3.1.1
odfpy==1.4.1
openpyxl==3.0.8
pandas==1.3.3
pycparser==2.20
PyJWT==2.1.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2021.1
PyYAML==5.4.1
requests==2.26.0
requests-oauthlib==1.3.0
six==1.16.0
SQLAlchemy==1.4.23
sqlparse==0.4.2
tablib==3.0.0
urllib3==1.26.6
xlrd==2.0.1
xlwt==1.3.0